requirements.txt
websocket_client==1.8.0
websockets==14.1
